+++ packages/address-picker/src/index.style.tsx
@@ -1,13 +1,11 @@
 import styled from 'styled-components';
 import { Info } from '@lendi-ui/icon';
 import { container } from '@lendi-ui/container';
-import { mt, pt } from '@lendi-ui/spacing';
+import { pt } from '@lendi-ui/spacing';
 
 export const Wrapper = styled.section`
   ${container()}
   max-width: 385px;
-  ${mt('xl')}
-  text-align: center;
 `;
 
 export const Alert = styled.div`